fix: make only row drag button draggable on mobile
