--- src/react/table-app/row-options/styles.css
@@ -1,3 +1,0 @@
-.NLT__row-options {
-    padding-left: var(--nlt-spacing--md);
-}
