+++ src/shared/dragging/types.ts
@@ -1,4 +1,4 @@
-export interface TableDataTransferItem {
+export interface DragData {
 	type: "row" | "column";
 	id: string;
 }
